refactor(editor): Move the notification contract to src/types/notification.ts
Pure relocation within `@n8n/composables`: the moved file's contents are
unchanged, and only the five import specifiers that referenced it move with
it. The subpath becomes `@n8n/composables/types/notification`, which resolves
through the existing `"./*"` exports map exactly as
`@n8n/frontend-utils/constants/sanitization` already does.

No consumer churn, so AC5 holds: `NotificationOptions` reaches its importers
through `@n8n/stores/notifications.store` and `@/Interface`, whose public
surfaces are untouched. No `editor-ui` file changes.
